move components to corresponding folder
